% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distribution.R
\name{alpha.edge.internal}
\alias{alpha.edge.internal}
\title{\code{alpha.edge.internal} yields the prior distribution for each situation
in the event tree associated with a particular variable.}
\usage{
alpha.edge.internal(level, tree, alpha)
}
\arguments{
\item{level}{"numeric" controls the number of descendent situations}

\item{tree}{"EventTree" a S4 object that represents an event tree.}

\item{alpha}{# TODO(taranti)}
}
\value{
# TODO(taranti)
}
\description{
\code{alpha.edge.internal} yields the prior distribution for each situation
in the event tree associated with a particular variable.
}
\examples{

}
