% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distribution.R
\name{AlphaEdgeInternal}
\alias{AlphaEdgeInternal}
\title{AlphaEdgeInternal}
\usage{
AlphaEdgeInternal(level, stratified.event.tree, alpha)
}
\arguments{
\item{level}{"numeric" controls the number of descendent situations}

\item{stratified.event.tree}{"Stratified.event.tree" a S4 object that
represents an event tree.}

\item{alpha}{numeric this represents a phantom sample used to initialize the
learning process}
}
\value{
numeric It is the hyperparameter of the Dirichlet prior distribution
 for each situation in the event tree associated with a particular variable.
}
\description{
\code{AlphaEdgeInternal} yields the prior distribution for each situation
in the event tree associated with a particular variable.
}

