% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{artificial.chds}
\alias{artificial.chds}
\title{test dataset - artificial.chds.}
\format{a data.frame with 1500 rows and 4 categorical variables. The variables
  names and values are compliant with CHDS, but the values are randomly filled.
   \describe{
   \item{Social}{High, Low}
   \item{Economic}{High, Low}
   \item{Event}{High, Average, Low}
   \item{Admission}{No, Yes}
   }}
\usage{
data(artificial.chds)
}
\description{
A dataset with dummy data, based on Child Health and Development Studies (CHDS).
}
\examples{
data(artificial.chds)
}
\keyword{datasets}
