% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_model.R
\docType{class}
\name{Ceg.model-class}
\alias{Ceg.model-class}
\title{Ceg.model S4 class}
\description{
\code{Ceg.model} is a S4 class which objects represents a Chain-Event Graph
(CEG) model, which is composed by a Staged Tree and its stages structure.
}
\section{Slots}{

\describe{
\item{\code{staged.tree}}{"Staged.tree"}

\item{\code{position}}{list  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}}

