% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_model.R
\name{CegGraph}
\alias{CegGraph}
\title{The function \code{CegGraph} produces a data structure corresponding a CEG
to be plotted using the package Rgraphviz.}
\usage{
CegGraph(event.tree, position, range.color = 1)
}
\arguments{
\item{event.tree}{"Event.tree"       an S4 object}

\item{position}{list       an object ceg.position.
# TODO(Collazo) DEFINIR MELHOR POSITION}

\item{range.color}{numeric    it chooses the color source.}
}
\value{
list
   \itemize{
   \item $node - node attributes
   \item $node$nodes (vector) - set of positions.
   \item node$variable (vector) - it identifies the variable asscoiated with
         each position.
   \item node$color (vector) - color of each position. All positions
         coincident with a stage are depicted in white.
   \item $edge - edge attributes
   \item $edge$edges (list) - set of list that emanates from each position.
   \item $edge$label (vector) - position labels.
   \item $edge$weight (vector) - edge weight.
   }
}
\description{
The function \code{CegGraph} produces a data structure corresponding a CEG
to be plotted using the package Rgraphviz.
}
\examples{

}
\seealso{
It uses \code{PositionVector}
}
