% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_model.R
\name{CegPosition}
\alias{CegPosition}
\title{CegPosition}
\usage{
CegPosition(stage, num.category, num.situation)
}
\arguments{
\item{stage}{(list) - stage structure associated with a particular level.}

\item{num.category}{(vector) - number of edges that unfolds from situations
in each level.}

\item{num.situation}{(vector) - number of situations associated with each
level.}
}
\value{
list of lists \itemize{
    \item First list level identifies a level 'l'.
    \item Second list level identifies the stage 'a' associated with
          level 'l'.
    \item The third list level identifies the positions associated with
          stage 'i' .
          }

 @seealso \code{PositionLevel},  \code{PositionVector},
 \code{PositionStage}, \code{PairwisePosition}
}
\description{
This function obtain the position structure associated with a particular
level of a CEG.
}

