% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{CheckAndCleanData}
\alias{CheckAndCleanData}
\title{CheckAndCleanData}
\usage{
CheckAndCleanData(data.frame)
}
\arguments{
\item{data.frame}{a data frame to be used to create stratified event/staged
trees}
}
\value{
data.frame with no void or NA values.
}
\description{
RemoveRowsWithNAandVoid remove all rows with NA and void ("") values data
from a data.frame
}
