% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{cont.table}
\alias{cont.table}
\title{This function calculates the contigency tables associated with each variable
in time-slices t_0 or t_k, k>=1.
#TODO(Collazo) definir melhor a funçao  cont.table.level - qual objetivo ou output}
\usage{
cont.table(data, tree)
}
\arguments{
\item{data}{(see function DCEG.AHC)}

\item{tree}{an object of type "EventTree"}
}
\value{
a list of matrices. Each matrix is the contigency table associate
        with a particular variable.
}
\description{
This function calculates the contigency tables associated with each variable
in time-slices t_0 or t_k, k>=1.
#TODO(Collazo) definir melhor a funçao  cont.table.level - qual objetivo ou output
}
\examples{

}
\seealso{
\code{\link{cont.table.level}} and \code{\link{cont.table.variable}}
}
