% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{ContingencyTable}
\alias{ContingencyTable}
\title{ContingencyTable}
\usage{
ContingencyTable(data, event.tree)
}
\arguments{
\item{data}{(see function DCEG.AHC) # TODO(Collazo) ????}

\item{event.tree}{an object of type "Event.tree"}
}
\value{
a list of matrices. Each matrix is the contigency table associate
        with a particular variable.
}
\description{
This function calculates the contigency tables associated with each variable
in time-slices t_0 or t_k, k>=1.
#TODO(Collazo) definir melhor a funçao  ContingencyTable/ContingencyTableLevel - qual objetivo ou output
}
\seealso{
\code{\link{ContingencyTableLevel}} and \code{\link{ContingencyTableVariable}}
}

