% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{ContingencyTableLevel}
\alias{ContingencyTableLevel}
\title{ContingencyTableLevel}
\usage{
ContingencyTableLevel(level, data, stratified.event.tree)
}
\arguments{
\item{level}{numeric - identifies the level in the infinite event tree.}

\item{data}{a data.frame structure  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{stratified.event.tree}{a Stratified.event.tree S4 object}
}
\value{
a matrix. Each row represents the contigency table of a situation
      k associate with a particular variable.
}
\description{
This function calculates the contigency tables associated with a specific
variable.
}

