% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{ContingencyTableVariable}
\alias{ContingencyTableVariable}
\title{ContingencyTableVariable}
\usage{
ContingencyTableVariable(variable, data, stratified.event.tree)
}
\arguments{
\item{variable}{numeric  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{data}{a data.frame structure  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{stratified.event.tree}{an object of type Stratified.event.tree}
}
\value{
a matrix. Each row represents the contigency table of a situation
         k associate with a particular variable.
}
\description{
This function calculates the contigency tables associated with a specific
variable according to a time-slice.
}

