% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{DataFrameChecker}
\alias{DataFrameChecker}
\title{DataFrameChecker}
\usage{
DataFrameChecker(data.frame)
}
\arguments{
\item{data.frame}{a data frame to be used to create stratified event/staged
trees}
}
\value{
data.frame with no void or NA values.

# TODO(Collazo)  dizer que quer usar o recurso, senao  taranti limpar o codigo.
}
\description{
DataFrameChecker verifies data behaviour and removes NULL and void data from
a provided data.frame
}
