% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirchlet_distribution.R
\docType{class}
\name{Dirchlet.distribution-class}
\alias{Dirchlet.distribution-class}
\title{Dirchlet.distribution}
\description{
Dirchlet.distribution
}
\section{Slots}{

\describe{
\item{\code{score}}{numeric.  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{\code{cluster}}{list.  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}}

