% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_probability.R
\docType{class}
\name{DistributionOfProbability-class}
\alias{DistributionOfProbability-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{score}}{numeric.}

\item{\code{cluster}}{list.}
}}

