% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{EdgeLabel}
\alias{EdgeLabel}
\title{EdgeLabel}
\usage{
EdgeLabel(num.variable, num.situation, label)
}
\arguments{
\item{num.variable}{numeric         - number of variables.}

\item{num.situation}{vector          - number of stages associated with
each variable.}

\item{label}{list of vectors - each component is a vector that
contains the event names associated with each variable.}
}
\value{
vector - edge labels
}
\description{
This function yields the edge labels. The edges are labeled accordingly the
original data provided.
}
