% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{EdgeSituation}
\alias{EdgeSituation}
\title{EdgeSituation}
\usage{
EdgeSituation(situation, start.situation, num.category)
}
\arguments{
\item{situation}{numeric  - it identifies the target situation whose
emanating edges are our interesting.}

\item{start.situation}{vector   - it identifies the situation that begins
a new level.}

\item{num.category}{vector   - it identifies the number of edges that
emanate from situations in each level.}
}
\value{
list of lists - each list component is a vector that represents the
 edges that emanate from a vertice.
}
\description{
EdgeSituation identifies the edges from a situation (node).
This function identifies the edges that emanate from a particular situation
in an EventTree.
}
