% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhaustive_model_search_algorithm.R
\docType{class}
\name{Exhaustive.model.search.algorithm-class}
\alias{Exhaustive.model.search.algorithm-class}
\title{Exhaustive.model.search.algorithm S4 Class}
\description{
Exhaustive.model.search.algorithm S4 Class
}
