% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_model_search_algorithm.R
\docType{class}
\name{Heuristic.model.search.algorithm-class}
\alias{Heuristic.model.search.algorithm-class}
\title{Heuristic.model.search.algorithm}
\description{
Heuristic.model.search.algorithm
}
