% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\docType{methods}
\name{initialize,Stratified.ceg.model-method}
\alias{initialize,Stratified.ceg.model-method}
\title{A simple "initialize" method. The real construction is performed by a
specific constructor. # TODO(taranti)  O CONSTRUTOR NAO FOI IMPLEMENTADO}
\usage{
\S4method{initialize}{Stratified.ceg.model}(.Object, stratified.staged.tree,
  position)
}
\arguments{
\item{position}{}

\item{Ceg.model}{}
}
\value{
Stratified.ceg.model
}
\description{
A simple "initialize" method. The real construction is performed by a
specific constructor. # TODO(taranti)  O CONSTRUTOR NAO FOI IMPLEMENTADO
}
\examples{
A simple "initialize" method. The real construction is performed by a
specific constructor. # TODO(taranti)  O CONSTRUTOR FOI IMPLEMENTADO NA FUNCAO CEG

 \\code{new("Stratified.ceg.model", stratified.stagedTree, position)}, however we recomend using
a specific function to construct the object # TODO(taranti)   LISTAR A FUNCAO
}
