% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\docType{methods}
\name{initialize,Stratified.event.tree-method}
\alias{initialize,Stratified.event.tree-method}
\title{Title}
\usage{
\S4method{initialize}{Stratified.event.tree}(.Object, num.variable,
  num.category, label.category, num.situation, num.slice, path, hyper.stage)
}
\description{
Title
}
