% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_staged_tree.R
\docType{methods}
\name{initialize,Stratified.staged.tree-method}
\alias{initialize,Stratified.staged.tree-method}
\title{Title}
\usage{
\S4method{initialize}{Stratified.staged.tree}(.Object,
  event.tree = "Stratified.event.tree", situation = "list",
  contingency.table = "list", stage.structure = "list",
  stage.probability = "list", prior.distribution = "list",
  posterior.distribution = "list", model.score = "numeric")
}
\description{
Title
}
