% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{KeepLexOrder}
\alias{KeepLexOrder}
\title{KeepLexOrder}
\usage{
KeepLexOrder(ref, order.vector, score.vector)
}
\arguments{
\item{ref}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{order.vector}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{score.vector}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}
\description{
This function keep a lexicographical order of a vector
}

