% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\name{LabelStage}
\alias{LabelStage}
\title{LabelStage}
\usage{
LabelStage(k, num.variable, num.situation, label.category, num.category)
}
\arguments{
\item{k}{numeric TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{num.variable}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{num.situation}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{label.category}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{num.category}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}
\value{
label   a vector TODO(Colazzo) Ampliar com tipo de dado e significado semantico
}
\description{
This function identifies the edges arriving at the target level for paths
that exist from the root node to each situation in the event tree that
are in levels greater than the target level.
}
\seealso{
\code{\link{TruncatedPath}}
}

