% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{MergeLabels}
\alias{MergeLabels}
\title{MergeLabels}
\usage{
MergeLabels(edge.list, edge, level)
}
\arguments{
\item{edge.list}{vector    list of positions that a children of a specific
position v1.}

\item{edge}{numeric    a particular children "edge" of a specific position v1}

\item{level}{vector    labels of classes corrresponding to the variable
associated with a position.}
}
\value{
list	 merged labels associated with a specific position v1.
}
\description{
Merge labels of multiple edges in order to plot them all.
}
\note{
This function mitigates a limitation from Rgraphviz, since it is not
possible to plot multiple edges between two nodes presenting the correct edge
label for each one.
The authors did not find a graphical package providing this capability.
Contributions are wellcomed.
}
