% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_distribution.R
\docType{class}
\name{Multinomial.distribution-class}
\alias{Multinomial.distribution-class}
\title{Multinomial.distribution}
\description{
Multinomial.distribution
}
\section{Slots}{

\describe{
\item{\code{score}}{numeric.}

\item{\code{cluster}}{list.}
}}

