% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_distribution.R
\docType{class}
\name{MultinomialDistribution-class}
\alias{MultinomialDistribution-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{score}}{numeric.}

\item{\code{cluster}}{list.}
}}

