% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{NaReorder}
\alias{NaReorder}
\title{NaReorder}
\usage{
NaReorder(x, order.score)
}
\arguments{
\item{x}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{order.score}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}
\description{
This function reorders a vector that had an element changed to NA.
}

