% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{NodeSet}
\alias{NodeSet}
\title{NodeSet}
\usage{
NodeSet(tree)
}
\arguments{
\item{tree}{an object event.tree. See files 2T.DCEG.ModelSearch.OAHC,
CEG.IP.ModelSearch and CEG.IP.ModelSearch.HS+AHC.}
}
\value{
vector
}
\description{
This function genereates the nodes of an event tree.
}

