% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\docType{class}
\name{OAHC-class}
\alias{OAHC-class}
\title{OAHC S4 Class}
\description{
@include heuristic_model_search_algorithm.R
}
\section{Slots}{

\describe{
\item{\code{score}}{numeric  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{\code{cluster}}{list  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}}

