% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{OAHC}
\alias{OAHC}
\title{This function calculates the best stage configuration of a hyperstage
  associated with a specific variable of time-slice t_0 or t_k, k>=1, using
  the oahc algorithm (oahc - Optimised Agglomerative Hierarchical Clustering)}
\usage{
OAHC(level, prior.distribution, contingency.table, tree)
}
\arguments{
\item{level}{numeric   - level under optimisation}

\item{prior.distribution}{(list of matrices) - see function
prior.distribution}

\item{contingency.table}{(list of matrices) - see function cont.table}

\item{tree}{an object 'event.tree' (see function event.tree)}
}
\description{
This function calculates the best stage configuration of a hyperstage
  associated with a specific variable of time-slice t_0 or t_k, k>=1, using
  the oahc algorithm (oahc - Optimised Agglomerative Hierarchical Clustering)
}
