% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{PairwiseScore}
\alias{PairwiseScore}
\title{PairwiseScore}
\usage{
PairwiseScore(stage.1, stage.2, vec.score, prior.distribution,
  contingency.table)
}
\arguments{
\item{contingency.table}{}
}
\description{
This function calculates the score asscoiated with a pair of merged stages.
}

