% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_model.R
\docType{methods}
\name{plot,Ceg.model,ANY-method}
\alias{plot,Ceg.model,ANY-method}
\title{Ceg.model Plotting}
\usage{
\S4method{plot}{Ceg.model,ANY}(x, y, ...)
}
\arguments{
\item{ceg.model}{a Ceg.model S4 object.}
}
\value{
the plot and also a pdf version is saved in the working directory.
}
\description{
Method to plot a chain event graph from a Ceg.model S4 object. The current
\code{ceg} package implementation depends on \code{Rgraphviz} package from
Bioconductor for plotting.
}
\examples{
plot(ceg.model)

}

