% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\docType{methods}
\name{plot,Stratified.ceg.model,ANY-method}
\alias{plot,Stratified.ceg.model,ANY-method}
\title{Stratified.ceg.model Plotting}
\usage{
\S4method{plot}{Stratified.ceg.model,ANY}(x)
}
\arguments{
\item{x}{Stratified.ceg.model S4 object.}
}
\value{
the plot and also a pdf version is saved in the working directory.
}
\description{
This Method is used to plot a chain event graph from a Stratified.ceg.model S4 object.
The current \code{ceg} package implementation depends on \code{Rgraphviz}
package from Bioconductor to draw the CEG graph.
}
\examples{
plot(scm)


}
