% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\docType{methods}
\name{plot,Stratified.event.tree,ANY-method}
\alias{plot,Stratified.event.tree,ANY-method}
\title{Title}
\usage{
\S4method{plot}{Stratified.event.tree,ANY}(x, y, ...)
}
\arguments{
\item{event.tree}{a Event.tree S4 object.}
}
\value{
the plot and also a pdf version is saved in the working directory.
}
\description{
Title
}
\note{
The current CEG package implementation depends on \code{Rgraphviz}
package from Bioconductor
}
\examples{

}
