% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\name{PositionLevel}
\alias{PositionLevel}
\title{PositionLevel}
\usage{
PositionLevel(stage.list, num.category, num.situation.next,
  pos.next.level = list())
}
\arguments{
\item{stage.list}{(list) - stage structure associated with a particular
variable.}

\item{num.category}{(vector) - number of edges that unfolds from each position
asscoiated with our target variable}

\item{num.situation.next}{(numeric) - number of situations associated with the variable
that follows our target variable in the event tree.}

\item{pos.next.level}{(list) - position structure associated with the
variable that follows our target variable in the event tree (see function PositionLevel)}
}
\value{
list of lists - The first list level identifies a stage 'i' and the
         second list level identifies the positions associated with this
         stage 'i'.
}
\description{
This function obtains the position structure associated with a particular
variable of a CEG.
}
\seealso{
\code{\link{PositionVector}}, \code{\link{PositionStage}} and
         \code{\link{PairwisePosition}}
}
