% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_model.R
\name{PositionVector}
\alias{PositionVector}
\title{\code{PositionVector} function rewrites a position structure associated with
      a particular level: from a list to a vector.}
\usage{
PositionVector(num.situation, pos.list)
}
\arguments{
\item{num.situation}{(numeric) - number of situation in a particular level.}

\item{pos.list}{(list) - stage structure of the level that follows the level
of our target position.}
}
\value{
vector
}
\description{
\code{PositionVector} function rewrites a position structure associated with
      a particular level: from a list to a vector.
}
\examples{

}
