% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distribution.R
\name{prior.variable}
\alias{prior.variable}
\title{The function \code{prior.variable}  yields the prior distribution for each
  situation in the event tree associated with a particular variable.}
\usage{
prior.variable(ref, alpha.edge)
}
\arguments{
\item{ref}{"numeric"  - variable order}

\item{alpha.edge}{"vector"   - the values of hyperparameters associated with
each edge of a situation corresponding to a particular variable.}
}
\value{
a matrix. The number of rows is equal to the number of situations
associated with a particular variable.
}
\description{
The function \code{prior.variable}  yields the prior distribution for each
  situation in the event tree associated with a particular variable.
}
\seealso{
\code{\link{Prior.distribution}} and
         \code{\link{alpha.edge.internal}}
}
