% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distribution.R
\name{PriorVariable}
\alias{PriorVariable}
\title{PriorVariable}
\usage{
PriorVariable(ref, alpha.edge)
}
\arguments{
\item{ref}{numeric  - It indicates the variable.}

\item{alpha.edge}{vector   - Dirichlet hyperparameter vector of a situation
associated with a particular variable.}
}
\value{
a matrix. Each row represents the Dirichlet hyperparameter vector of a 
                    situation associated with a particular variable in the event tree.
}
\description{
The function \code{PriorVariable}  yields the prior distributions for
  all situations associated with a particular variable in the event tree.
}
\seealso{
\code{Prior.distribution} and
         \code{AlphaEdgeInternal}
}
