% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{scm}
\alias{scm}
\title{test stratified ceg model}
\format{a Stratified.ceg.model S4 object}
\usage{
data(scm)
}
\description{
A Stratified.ceg.model S4 object, generated using the command
\code{scm <- Stratified.ceg.model(sst)}
}
\examples{
data(scm)
}
\keyword{datasets}
