% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{set.manual}
\alias{set.manual}
\title{test stratified event tree (manualy constructed)}
\format{a Stratified.event.tree S4 object}
\usage{
data(set)
}
\description{
A Stratified.event.tree S4 object, generated using manual input. \cr
See Stratified.event.tree documentation examples.
}
\examples{
data(set.manual)
}
\keyword{datasets}
