% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{set}
\alias{set}
\title{test stratified event tree}
\format{a Stratified.event.tree S4 object}
\usage{
data(set)
}
\description{
A Stratified.event.tree S4 object, generated using the command
\code{set <- Stratified.event.tree(artificial.chds)}
}
\examples{
data(set)
}
\keyword{datasets}
