% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{SingleReorder}
\alias{SingleReorder}
\title{SingleReorder}
\usage{
SingleReorder(x, order.score, score, na.count)
}
\arguments{
\item{na.count}{}
}
\description{
This function reorders a vector that had an element changed.
}
