% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{SingleScore}
\alias{SingleScore}
\title{SingleScore}
\usage{
SingleScore(stage, prior.distribution, contingency.table)
}
\arguments{
\item{stage}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{prior.distribution}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{contingency.table}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}
\description{
This function calculates the score associated with a particular level.
}

