% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sst}
\alias{sst}
\title{test stratified staged tree}
\format{a Stratified.staged.tree S4 object}
\usage{
data(sst)
}
\description{
A Stratified.staged.tree S4 object, generated using the command
\code{sst <- Stratified.staged.tree(artificial.chds)}
}
\examples{
data(sst)
}
\keyword{datasets}
