% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged_tree.R
\docType{class}
\name{Staged.tree-class}
\alias{Staged.tree-class}
\title{Title # TODO(taranti)}
\description{
Title # TODO(taranti)
}
\section{Slots}{

\describe{
\item{\code{event.tree}}{Event.tree.}

\item{\code{prior.distribution}}{list.}

\item{\code{contingency.table}}{list.}

\item{\code{stage.structure}}{list.}

\item{\code{model.core}}{numeric.}
}}

