% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\docType{class}
\name{Stratified.ceg.model-class}
\alias{Stratified.ceg.model-class}
\title{Stratified.ceg.model}
\description{
The \code{Stratified.ceg.model} is a S4 class that extends \code{Ceg.model}.
 The object represents a CEG model derived from its supporting Stratified.staged.tree
 using some graphical transformation rules.
}
