% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\name{Stratified.ceg.model}
\alias{Stratified.ceg.model}
\title{StratifiedCegModel constructor.}
\usage{
Stratified.ceg.model(Stratified.staged.tree, stages)
}
\arguments{
\item{stages}{list         TODO(Collazo)  describe}

\item{StratifiedStagedTree}{S4 object    TODO(Collazo)  describe}
}
\value{
a StratifiedCegModel S4 object.
}
\description{
S3 function to friendly construct S4  StratifiedCegModel.
}
\details{
The \code{stratified.ceg.model} returns a S4 StratifiedCegModel.
}
\examples{
TODO(taranti) depois de resolvida a confusao entre stage.structure e position

}
\seealso{
\code{\link{StratifiedCegModel}}
}
