% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\docType{class}
\name{Stratified.event.tree-class}
\alias{Stratified.event.tree-class}
\title{Stratified.event.tree S4 Class}
\description{
An event tree is called stratified if the set of events that unfold from all situations,
which are at the same distance of edges from the initial situation, are identical.
}
