% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\name{StratifiedCegPosition}
\alias{StratifiedCegPosition}
\title{StratifiedCegPosition}
\usage{
StratifiedCegPosition(stage, num.category, num.situation)
}
\arguments{
\item{stage}{(list) - stage structure associated with a particular variable.}

\item{num.category}{(vector) - number of edges that unfold from stages
associated with a particular variable.}

\item{num.situation}{(vector) - number of situations associated with each
variable.}
}
\value{
list of lists \itemize{
    \item First list level identifies a variable 'v'.
    \item Second list level identifies a stage 'a' associated with
          a variable 'v'.
    \item The third list level identifies the positions associated with
          a stage 'a' .
          }

 @seealso \code{PositionLevel},  \code{PositionVector},
 \code{PositionStage}, \code{PairwisePosition}
}
\description{
This function obtains the position structure associated with
a stratified CEG.
}
