% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\name{StratifiedEventTreeGraph}
\alias{StratifiedEventTreeGraph}
\title{Title}
\usage{
StratifiedEventTreeGraph(event.tree)
}
\arguments{
\item{event.tree}{"Event.tree" S4 object

 @return list contains a data structure to be used in a plot}
}
\description{
Title
}
