% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\name{TruncatedPath}
\alias{TruncatedPath}
\title{TruncatedPath}
\usage{
TruncatedPath(ref, k, var, num.category, num.situation, label.category)
}
\arguments{
\item{ref}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{k}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{var}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{num.category}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{num.situation}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{label.category}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}
\description{
This function identifies the edges arriving at the target level for all
paths that exist from the root node to each situation in the event tree.
}

