% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\docType{class}
\name{Variable-class}
\alias{Variable-class}
\title{Variable S4 Class}
\description{
Variable S4 class contains two slots with the Variable name and a list
of Categories. It is used to construct Stratified.vent.tree objects.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{categories}}{list of Category S4 objects.}
}}

