% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{Variable}
\alias{Variable}
\title{Variable(name,categories)}
\usage{
Variable(name, categories)
}
\arguments{
\item{name}{character, the Variable name}

\item{categories}{a list of S4 Category objects.}
}
\value{
a \code{\link{Variable}} S4 object
}
\description{
Variable(name,categories) is a function that act as constructor to Variable
S4 object. Variable S4 class contains two slots with the Variable name and a
list of Categories. It is used to construct Stratified.vent.tree objects.
}
\examples{
var <- Variable("variable.name", list(Category("cat1"), Category("cat2"),
Category("cat3")))


}
