% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distribution.R
\name{AlphaEdgeInternal}
\alias{AlphaEdgeInternal}
\title{AlphaEdgeInternal}
\usage{
AlphaEdgeInternal(level, stratified.event.tree, alpha)
}
\arguments{
\item{level}{numeric - It indicates the level in the event tree.}

\item{stratified.event.tree}{Stratified.event.tree - S4 object that represents an event tree.}

\item{alpha}{numeric - It plays a role of a phantom sample to construct 
the prior probability distribution of a situation associated
with a particular variable in the event tree.}
}
\value{
"vector"  - Dirichlet hyperparameter vector of a situation
                    associated with a particular variable.
}
\description{
\code{AlphaEdgeInternal} yields a possible objective prior distribution for each situation
associated with a particular variable in the event tree.
}
