% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distribution.R
\name{AlphaEdgeInternal}
\alias{AlphaEdgeInternal}
\title{AlphaEdgeInternal}
\usage{
AlphaEdgeInternal(level, tree, alpha)
}
\arguments{
\item{level}{"numeric" controls the number of descendent situations}

\item{tree}{"EventTree" a S4 object that represents an event tree.}

\item{alpha}{# TODO(Collazo)}
}
\value{
# TODO(Collazo) descrever retorno, se necessario.
}
\description{
\code{AlphaEdgeInternal} yields the prior distribution for each situation
in the event tree associated with a particular variable.
}

