% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\docType{class}
\name{Category-class}
\alias{Category-class}
\title{Category S4 Class}
\description{
Category S4 class contains a single slot with the category label. It is used
to construct Stratified.event.tree objects.
}
\section{Slots}{

\describe{
\item{\code{label}}{character}
}}

