% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{Category}
\alias{Category}
\title{Category(label)}
\usage{
Category(label)
}
\arguments{
\item{label}{caracter, the category name}
}
\value{
a \code{\link{Category}} S4 object
}
\description{
Category(label) is a function that act as constructor to Category S4 object.
Category S4 class contains a single slot with the category labels. It is used
to construct S4 Variable objects, which, in turn, aim at being parameters in
Stratified.event.tree objects manual constructions.
}
\examples{
cat <- Category("category.name")

Category("category.name2")


}
