% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_model.R
\docType{class}
\name{Ceg.model-class}
\alias{Ceg.model-class}
\title{Ceg.model S4 class}
\description{
\code{Ceg.model} is a S4 class whose objects represent a Chain-Event Graph
(CEG) model, which is composed by a Staged Tree object and its corresponding staged structure.
}
\section{Slots}{

\describe{
\item{\code{staged.tree}}{Staged.tree S4 object}

\item{\code{position}}{list}
}}

