% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{ContingencyTable}
\alias{ContingencyTable}
\title{ContingencyTable}
\usage{
ContingencyTable(data, stratified.event.tree)
}
\arguments{
\item{data}{data.frame whose columns depict variables and rows correspond
to units that are observed in the system}

\item{stratified.event.tree}{Stratified.event.tree S4 object}
}
\value{
a list of matrices that represent the contigency tables associated with each variable in the event tree. 
        The matrix corresponding to a particular variable presents the counts of each combination of the categories  
     of the variables that precede it in the event tree according to its categories. The combinations of the categories
     of the upstream variables are displayed on the rows and represent the situations associated with the target variable.
     The categories of the target variable are represented on the columns and corresponds to each event that can unfold 
     from a situation associated with the target variable.
}
\description{
This function creates the contigency tables associated with each variable in the event tree.
}
