% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{ContingencyTable}
\alias{ContingencyTable}
\title{ContingencyTable}
\usage{
ContingencyTable(data, stratified.event.tree)
}
\arguments{
\item{data}{a data.frame structure  TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{stratified.event.tree}{an object of type "Stratified.event.tree"}
}
\value{
a list of matrices. Each matrix is the contigency table associate
        with a particular variable.
}
\description{
This function calculates the contigency tables associated with each variable.
This enables us to calculate the number of units that visit each situation
in an Stratified.event.tree.
This function is used to generate a stratified.staged.tree when a well
behavioured data.frame is provided as an argument to Stratified.staged.tree()
constructor.
}
\seealso{
\code{\link{ContingencyTableLevel}} and \code{\link{ContingencyTableVariable}}
}

