% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{ContingencyTableLevel}
\alias{ContingencyTableLevel}
\title{ContingencyTableLevel}
\usage{
ContingencyTableLevel(level, data, event.tree)
}
\arguments{
\item{level}{numeric - identifies the level in the infinite event tree.}

\item{data}{data #TODO(Collazo) definir o que é data (see function DCEG.AHC)}

\item{tree}{an "Event.tree" S4 object}
}
\value{
a matrix. Each row represents the contigency table of a situation
      k associate with a particular variable in time-slice t_0 or t_k, k>=1.
}
\description{
This function calculates the contigency tables associated with a specific
variable in time-slices t_0 or t_k, k>=1.
}
