% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{ContingencyTableVariable}
\alias{ContingencyTableVariable}
\title{ContingencyTableVariable}
\usage{
ContingencyTableVariable(variable, data, tree)
}
\arguments{
\item{variable}{numeric}

\item{data}{#TODO(Collazo) definir o que é data}

\item{tree}{an object of type 'event.tree' (see function event.tree)}
}
\value{
a matrix. Each row represents the contigency table of a situation
         k associate with a particular variable.
}
\description{
This function calculates the contigency tables associated with a specific
variable according to a time-slice.
}
