% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{DataFrameChecker}
\alias{DataFrameChecker}
\title{verifies data complience}
\usage{
DataFrameChecker(data.frame)
}
\arguments{
\item{data.frame}{a data frame to be used to create stratified event/staged
trees}
}
\value{
a data.frame with no void or NA values.
}
\description{
verifies data behaviour and removes NULL and void data
}
\details{
This function verifies data behaviour and removes NA and void data.
}
\examples{
TODO(taranti)

}
