% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirchlet_mpnl_distribution.R
\docType{class}
\name{Dirchlet.MPNL.distribution-class}
\alias{Dirchlet.MPNL.distribution-class}
\title{Dirchlet.MPNL.distribution}
\description{
Dirchlet.MPNL.distribution
}
\section{Slots}{

\describe{
\item{\code{score}}{numeric.}

\item{\code{cluster}}{list.}
}}

