% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_probability.R
\docType{class}
\name{Distribution.of.probability-class}
\alias{Distribution.of.probability-class}
\title{Distribution.of.probability S4 Class}
\description{
Distribution.of.probability S4 Class
}
\section{Slots}{

\describe{
\item{\code{score}}{numeric.}

\item{\code{cluster}}{list.}
}}

