% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{EdgeList}
\alias{EdgeList}
\title{EdgeList}
\usage{
EdgeList(stratified.event.tree, node)
}
\arguments{
\item{stratified.event.tree}{Stratified.event.tree S4 object}

\item{node}{(vector) - an object generated by the function node.list}
}
\value{
list of lists - each list component is a vector that represents the
edges that emanate from a vertice.
}
\description{
Function EdgeList genereates the list of edges of an event tree.
}
