% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{KeepLexOrder}
\alias{KeepLexOrder}
\title{KeepLexOrder}
\usage{
KeepLexOrder(ref, order.vector, score.vector)
}
\arguments{
\item{score.vector}{}
}
\description{
This function keep a lexicographical order of a vector
}
