% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\name{LabelStage}
\alias{LabelStage}
\title{LabelStage}
\usage{
LabelStage(k, num.variable, num.situation, label.category, num.category)
}
\arguments{
\item{k}{numeric}

\item{num.variable}{numeric}

\item{num.situation}{numeric}

\item{label.category}{list}

\item{num.category}{list

 @return   label   a vector}
}
\description{
This function identifies the edges arriving at the target level for paths
that exist from the root node to each situation in the event tree that
are in levels greater than the target level.
}
\seealso{
\code{\link{TruncatedPath}}
}
