% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{ListToVector}
\alias{ListToVector}
\title{ListToVector}
\usage{
ListToVector(x, n)
}
\arguments{
\item{x}{list of vectors TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{n}{numeric TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}
\value{
vector
}
\description{
This function change a list of vectors in a vector.
}

