% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{ListToVector}
\alias{ListToVector}
\title{change a list of vectors in a vector}
\usage{
ListToVector(x, n)
}
\arguments{
\item{x}{list of vectors}

\item{n}{numeric}
}
\value{
vector
}
\description{
This function change a list of vectors in a vector.
}
\details{
This function change a list of vectors in a vector.
}
\examples{


}
