% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{NodeLabel}
\alias{NodeLabel}
\title{yields the node labels.}
\usage{
NodeLabel(num.variable, num.situation, num.category, label)
}
\arguments{
\item{num.variable}{numeric - number of variables.}

\item{num.situation}{vector - number of stages associated with each
variable.}

\item{num.category}{vector - it identifies the number of edges that emanate
from situations in each level.}

\item{label}{list of vectors - each component is a vector that
contains the event names associated with each variable.}
}
\value{
vector - node labels
}
\description{
This function yields the node labels.
}
\details{
The nodes are labeled accordingly, do indicate diferente positions.
TODO(taranti)  verify details.
}
