% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{OAHC}
\alias{OAHC}
\title{OAHC Constructor}
\usage{
OAHC(level, prior.distribution, contingency.table, tree)
}
\arguments{
\item{level}{numeric   - level under optimisation}

\item{prior.distribution}{(list of matrices) - see function
\code{prior.distribution}}

\item{contingency.table}{(list of matrices) - see function ContingencyTable}

\item{tree}{an object 'Event.tree'}
}
\value{
a OAHC S4 object
}
\description{
This function calculates the best stage configuration of a hyperstage
  associated with a specific variable of time-slice t_0 or t_k, k>=1, using
  the oahc algorithm (oahc - Optimised Agglomerative Hierarchical Clustering)
}
\seealso{
SingleScore,  PairwiseScore, SingleReorder, NaReorder, KeepLexOrder
}
