% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\name{PairwisePosition}
\alias{PairwisePosition}
\title{PairwisePosition}
\usage{
PairwisePosition(pair.situation, num.category, pos.next.level)
}
\arguments{
\item{pair.situation}{(vector) - situations to be analysed}

\item{num.category}{(numeric) - number of edges that unfolds from the
situations}

\item{pos.next.level}{(vector) - It identifies the positions for all
situations in the next level.}
}
\value{
boolean
}
\description{
The \code{PairwisePosition } function identifies if two situations are in
the same position given that they are in the same stage.
}

