% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_model.R
\docType{methods}
\name{plot,Ceg.model,ANY-method}
\alias{plot,Ceg.model,ANY-method}
\title{Title}
\usage{
\S4method{plot}{Ceg.model,ANY}(x, y, ...)
}
\arguments{
\item{Ceg.model}{a CegModel S4 object.}
}
\value{
the plot and also a pdf version is saved in the working directory.
}
\description{
Title
}
\note{
The current CEG package implementation depends on \code{Rgraphviz}
package from Bioconductor
}
\examples{

}
