% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_staged_tree.R
\docType{methods}
\name{plot,Stratified.staged.tree,ANY-method}
\alias{plot,Stratified.staged.tree,ANY-method}
\title{Title}
\usage{
\S4method{plot}{Stratified.staged.tree,ANY}(x, y, ...)
}
\arguments{
\item{Staged.tree}{}
}
\description{
Title
}
