% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_model.R
\name{PositionStage}
\alias{PositionStage}
\title{\code{PositionStage} function yields the position structure associated with
a particular
stage of a CEG.}
\usage{
PositionStage(stage.vector, num.category, pos.next.level)
}
\arguments{
\item{stage.vector}{(vector) - a set of situations that constitute a
particular stage}

\item{num.category}{(numeric) - number of edges that unfolds from the
situations}

\item{pos.next.level}{(vector) - It identifies the positions for all
situations in the next level.#'}
}
\value{
list of vector - Each vector identifies a position.
}
\description{
\code{PositionStage} function yields the position structure associated with
a particular
stage of a CEG.
}
\examples{

}
\seealso{
\code{\link{PairwisePosition}}
}
