% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\name{PositionVector}
\alias{PositionVector}
\title{\code{PositionVector} function rewrites a position structure associated with
      a particular variable: from a list to a vector.}
\usage{
PositionVector(num.situation, pos.list)
}
\arguments{
\item{num.situation}{(numeric) - number of situation associated with a particular variable.}

\item{pos.list}{(list) - stage structure associated with a particular variable that follows
the variable associated with our target position.}
}
\value{
vector
}
\description{
\code{PositionVector} function rewrites a position structure associated with
      a particular variable: from a list to a vector.
}
