% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distribution.R
\name{PriorDistribution}
\alias{PriorDistribution}
\title{PriorDistribution}
\usage{
PriorDistribution(tree, alpha)
}
\arguments{
\item{tree}{"EventTree" a S4 object that represents an event tree.}

\item{alpha}{# TODO(Collazo)  -- see function DCEG.AHC}
}
\value{
prior is a list of matrices. Each matrix is a collection of
vectors that correspond to a prior for each situation associated with
a particular variable.
}
\description{
\code{PriorDistribution} initialises the prior distributions under the
conservative and uniform assumptions for the hyperparameter 'alpha' over
the event tree.
}
\seealso{
\code{\link{PriorVariable}}
}
