% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{SingleReorder}
\alias{SingleReorder}
\title{SingleReorder}
\usage{
SingleReorder(x, order.score, score, na.count)
}
\arguments{
\item{x}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{order.score}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{score}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}

\item{na.count}{TODO(Colazzo) Ampliar com tipo de dado e significado semantico}
}
\description{
This function reorders a vector that had an element changed.
}

