% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oahc.R
\name{SingleScore}
\alias{SingleScore}
\title{SingleScore}
\usage{
SingleScore(stage, prior.distribution, contingency.table)
}
\arguments{
\item{contingency.table}{}
}
\description{
This function calculates the score associated with a particular level.
}

