% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged_tree.R
\docType{class}
\name{Staged.tree-class}
\alias{Staged.tree-class}
\title{Staged.tree}
\description{
A staged tree is an event tree embellished with colours using a probabilistic measure. 
 Two situations are said to be in the same stage if they have equivalent probabilistic space and
 identical conditional probabilities. Each stage is associated with a different colour.
}
\section{Slots}{

\describe{
\item{\code{event.tree}}{Event.tree.}
}}

