% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\docType{class}
\name{Stratified.ceg.model-class}
\alias{Stratified.ceg.model-class}
\title{The \code{Stratified.ceg.model} is a S4 class thar extends \code{\link{Ceg.model}}. Its
 objects represents a CEG model derived from a Stratified.staged.tree.}
\value{
a "Stratified.ceg.model" S4 object.
}
\description{
The \code{Stratified.ceg.model} is a S4 class thar extends \code{\link{Ceg.model}}. Its
 objects represents a CEG model derived from a Stratified.staged.tree.
}
