% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\name{Stratified.ceg.model}
\alias{Stratified.ceg.model}
\title{Stratified.ceg.model constructor.}
\usage{
Stratified.ceg.model(stratified.staged.tree)
}
\arguments{
\item{stratified.staged.tree}{Stratified.staged.tree S4 object    A staged tree is called stratified if
its supporting event tree is stratified and all vertices which are in the same stage
are also at the same distance of edges from the root.}
}
\value{
a Stratified.ceg.model S4 object.
}
\description{
S3 function to friendly construct S4  Stratified.ceg.model.
}
\examples{
scm <- Stratified.ceg.model(sst)

}
