% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_ceg_model.R
\name{Stratified.ceg.model}
\alias{Stratified.ceg.model}
\title{Stratified.ceg.model constructor.}
\usage{
Stratified.ceg.model(stratified.staged.tree)
}
\arguments{
\item{Stratified.staged.tree}{S4 object    TODO(Collazo)  describe}
}
\value{
a Stratified.ceg.model S4 object.
}
\description{
S3 function to friendly construct S4  Stratified.ceg.model.
}
\examples{
\dontrun{stratified.ceg.model <- Stratified.ceg.model(stratified.staged.tree, position)}

}

