% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\docType{class}
\name{Stratified.event.tree-class}
\alias{Stratified.event.tree-class}
\title{Stratified.event.tree S4 Class}
\description{
# TODO(Collazo) se considerar necessario, explicar o que é Stratified.event.tree
}
