% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\docType{methods}
\name{Stratified.event.tree}
\alias{Stratified.event.tree}
\alias{Stratified.event.tree,missing-method}
\alias{Stratified.event.tree,ANY-method}
\alias{Stratified.event.tree,list-method}
\title{Stratified.event.tree}
\usage{
Stratified.event.tree(x, ...)

\S4method{Stratified.event.tree}{missing}(x)

\S4method{Stratified.event.tree}{ANY}(x, ...)

\S4method{Stratified.event.tree}{list}(x = "list")
}
\arguments{
\item{Arguments}{(missing) \cr
A call to \code{Stratified.event.tree( )} with no parameters will return
an error message for missing argument.}

\item{Arguments}{(ANY) \cr
A call to \code{Stratified.event.tree(x, ...)}, x not being a data.frame or
a list, will return an error message.}

\item{Arguments}{(list) , list of Variable S4 objects, in the expected order
of plotting. \cr
TODO(Collazo) VRF texto}
}
\value{
a Stratified.event.tree S4 object
}
\description{
Constructor method to Stratified.event.tree S4 objects. It accepts different
sets for parameters types.
}
