% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\name{Stratified.event.tree}
\alias{Stratified.event.tree}
\title{Title}
\usage{
Stratified.event.tree(x, ...)
}
\arguments{
\item{...}{Extra named arguments passed to FUN}

\item{object}{An object}

\item{data}{Numeric vector or data.frame}

\item{Fun}{Function. Default function is \code{sum}}
}
\description{
Title
}
