% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_staged_tree.R
\docType{class}
\name{Stratified.staged.tree-class}
\alias{Stratified.staged.tree-class}
\title{Title  # TODO(taranti)}
\description{
Title  # TODO(taranti)
}
\section{Slots}{

\describe{
\item{\code{event.tree}}{Stratified.event.tree}
}}

