% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_staged_tree.R
\docType{class}
\name{Stratified.staged.tree-class}
\alias{Stratified.staged.tree-class}
\title{Stratified.staged.tree}
\description{
#TODO(Collazo) Definir o que é stratified staged.tree
}
\section{Slots}{

\describe{
\item{\code{event.tree}}{Stratified.event.tree.}

\item{\code{situation}}{list.}

\item{\code{contingency.table}}{list.}

\item{\code{stage.structure}}{list.}

\item{\code{stage.probability}}{list.}

\item{\code{prior.distribution}}{list.}

\item{\code{posterior.distribution}}{list.}

\item{\code{model.score}}{numeric.}
}}

