% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_staged_tree.R
\docType{methods}
\name{Stratified.staged.tree}
\alias{Stratified.staged.tree}
\alias{Stratified.staged.tree,missing,ANY,ANY-method}
\alias{Stratified.staged.tree,ANY,ANY,ANY-method}
\alias{Stratified.staged.tree,data.frame,numeric,numeric-method}
\alias{Stratified.staged.tree,Stratified.event.tree,list,ANY-method}
\title{Stratified.staged.tree}
\usage{
Stratified.staged.tree(x, y, z, ...)

\S4method{Stratified.staged.tree}{missing,ANY,ANY}(x, y, z, ...)

\S4method{Stratified.staged.tree}{ANY,ANY,ANY}(x, y, z, ...)


  \S4method{Stratified.staged.tree}{data.frame,numeric,numeric}(x = "data.frame",
  y = 1L, z = 0L)


  \S4method{Stratified.staged.tree}{Stratified.event.tree,list,ANY}(x = "Stratified.event.tree",
  y = "list")
}
\arguments{
\item{Arguments}{(missing) \cr
A call to \code{Stratified.staged.tree( )} with no parameters will return
an error message for missing argument.}

\item{Arguments}{(ANY) \cr
A call to \code{Stratified.staged.tree(x, ...)}, x not being a data.frame or
a Event.tree, will return an error message.}

\item{Arguments}{(data.frame, numeric, numeric) , where data.frame is a well
behavioured data set, and the numeric values represent the alpha and
the variable order, respectively.\cr
The implementation admits providing the three arguments, or the first two, or
 even only the data.frame.\cr
 The default variable order is as in the data.frame and the default alpha is
 1L.}

\item{Arguments}{(Stratified.event.tree, list) , where datalist.frame is the
stage.structure to be aplied on the provided Stratified.event.tree. \cr
To construct the aforesaid list one must generate the Stratified.event.tree
graph and use the number of each presented node. These numbers will indicate
the nodes that share the same position in a stage \cr
TODO(Collazo) VRF texto}
}
\value{
a Stratified.staged.tree S4 object
}
\description{
Constructor method to Stratified.staged.tree S4 objects. It accepts different
sets for parameters types.
}
