% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_staged_tree.R
\name{Stratified.staged.tree}
\alias{Stratified.staged.tree}
\title{Title}
\usage{
Stratified.staged.tree(x, y, z, ...)
}
\arguments{
\item{...}{Extra named arguments passed to FUN}

\item{object}{An object}

\item{data}{Numeric vector or data.frame}

\item{Fun}{Function. Default function is \code{sum}}
}
\description{
Title
}
