% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\name{StratifiedEventTreeGraph}
\alias{StratifiedEventTreeGraph}
\title{StratifiedEventTreeGraph}
\usage{
StratifiedEventTreeGraph(event.tree)
}
\arguments{
\item{event.tree}{"Event.tree" S4 object

 @return list with a data structure that is adequate to plot an event tree}
}
\description{
StratifiedEventTreeGraph
}
