% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_event_tree.R
\name{TruncatedPath}
\alias{TruncatedPath}
\title{TruncatedPath}
\usage{
TruncatedPath(ref, k, var, num.category, num.situation, label.category)
}
\arguments{
\item{label.category}{}
}
\description{
This function identifies the edges arriving at the target level for all
paths that exist from the root node to each situation in the event tree.
}
