% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_funtions.R
\name{Z.NA.Keep}
\alias{Z.NA.Keep}
\title{Z.NA.Keep}
\usage{
Z.NA.Keep(df, col = NULL, n = 0)
}
\arguments{
\item{n}{}
}
\description{
Remove void and NA values from dataframe.
The code is strongly based on the one provide by Jerry T. in
 \link{http://stackoverflow.com/questions/4862178/remove-rows-with-nas-in-data-frame}.
}
\note{
credit: http://stackoverflow.com/a/30461945/2292993
}
\seealso{
\link{http://stackoverflow.com/a/30461945/2292993}.
}
